chore: Update snippet text and search params for debugger search menu (#7249)
