+++ app/client/src/components/editorComponents/Debugger/DebuggerLogs.tsx
@@ -5,7 +5,7 @@ import { Severity } from "entities/AppsmithConsole";
 import FilterHeader from "./FilterHeader";
 import { BlankState } from "./helpers";
 import LogItem, { getLogItemProps } from "./LogItem";
-import { usePagination, useFilteredLogs } from "./hooks";
+import { usePagination, useFilteredLogs } from "./hooks/debuggerHooks";
 import { createMessage, NO_LOGS } from "constants/messages";
 import { useSelector } from "react-redux";
 import { getCurrentUser } from "selectors/usersSelectors";
