+++ app/client/src/constants/messages.ts
@@ -364,7 +364,7 @@ export const TROUBLESHOOT_ISSUE = () => "Troubleshoot issue";
 export const DEBUGGER_SEARCH_GOOGLE = () => "Ask Google";
 export const DEBUGGER_COPY_MESSAGE = () => "Copy";
 export const DEBUGGER_OPEN_DOCUMENTATION = () => "Open Documentation";
-export const DEBUGGER_SEARCH_SNIPPET = () => "Trigger Snippet";
+export const DEBUGGER_SEARCH_SNIPPET = () => "Browse Code Snippets";
 export const DEBUGGER_APPSMITH_SUPPORT = () => "Get Appsmith Support";
 
 //action creator menu
